fixed casts

+++ src/main/java/eu/qped/racket/functions/numbers/Negative.java
@@ -22,7 +22,7 @@ public class Negative extends Expression {
         for (Class<?> clazz : opNum.arrayList) {
             count++;
             if (list.get(0) instanceof Number || clazz.isInstance(list.get(0).getParts().get(0))) {
-                value = (int) list.get(0).evaluate(this);       //Because Racket only accepts Integers in even?
+                value = (int)(float) list.get(0).evaluate(this);       //Because Racket only accepts Integers in even?
                 break;
             } else {
                 if (opNum.arrayList.size() == count) {
